refactor(files): change name for all files from camelCase to kebab-case
