+++ src/common/entities/base-entity.entity.ts
